[REFACTOR][TOPI] Migrate topi container signatures + stray Integer/Bool callers
- Change ffi::Array<Integer> to ffi::Array<int64_t> in topi headers
  (strided_slice, reduction, transform, nn, utils, nn/group_norm,
  nn/instance_norm, nn/layer_norm, nn/rms_norm) and corresponding
  sources; update internal usages (.IntValue(), .defined(), ->value,
  ->dtype) to plain int64_t arithmetic
- Migrate ShardingNode::sharding_dim and CalculateConstantBytes/
  CalculateWorkspaceBytes signatures from Integer to int64_t
- Migrate codegen_c.h constants_byte_alignment_ field from Integer
  to int64_t
- Replace Downcast<Integer/Bool> with Downcast<IntImm> in schedule
  primitives, metaschedule helpers, and misc transform passes
- Convert remaining Integer/Bool local variables and function return
  types (clml codegen, make_packed_api, infer_layout_utils, etc.)
  to IntImm or plain int64_t
- Update stale doc-comment "Type: Integer/Bool" annotations in
  function.h and tirx/function.h to "Type: IntImm"
- Fix test files: nested_msg_test.cc (NestedMsg<IntImm>), ir_functor_test.cc,
  arith_simplify_test.cc

+++ include/tvm/topi/nn/layer_norm.h
@@ -49,7 +49,7 @@ using namespace tvm::te;
  * \return The normalized tensor, with the same shape as data.
  */
 inline Tensor layer_norm(const Tensor& data, const Tensor& gamma, const Tensor& beta,
-                         const ffi::Array<Integer>& axis, double epsilon,
+                         const ffi::Array<int64_t>& axis, double epsilon,
                          std::string name = "T_layer_norm", std::string tag = kInjective) {
   const auto& data_type = data->dtype;
   const auto& gamma_type = gamma.defined() ? gamma->dtype : data_type;
